app/requirements.txt
cassandra-driver
venv-pack
pyspark
